artifacts/stock_analysis_NVDA-TSLA_20251015_055626/code/v001_critic_stock_service.py
# Auto-generated StockService v1
# Generated at: 2025-10-15T05:57:06.248596

import yfinance as yf
import pandas as pd
import numpy as np
from typing import Dict, List

class StockServiceV1:
    """Stock service with evolved capabilities"""
    
    def __init__(self):
        self.version = 1
        self.capabilities = {
        "prices": true,
        "moving_avg": false,
        "rsi": false,
        "volatility": false,
        "correlation": false,
        "volume": false
}
    
    def get_stock_prices(self, symbols: List[str]) -> Dict[str, pd.DataFrame]:
        """Fetch stock data with enhanced features"""
        # Enabled capabilities: prices
        
        data = {}
        for symbol in symbols:
            ticker = yf.Ticker(symbol)
            df = ticker.history(start=f"{datetime.now().year}-01-01")

            data[symbol] = df
        return data
