requirements.txt
ag2[openai]
streamlit
yfinance
matplotlib
pandas
python-dotenv
